Fix --nosummary flag
The `--nosummary` flag was inverted, which cause it to always be true.

Rather than just flip `store_false` to `store_true`, I opted to refactor the
internal name from `nosummary` to `showSummary`.  This avoids double-negatives
("Are we not not showing the summary?"), making the logic easier to follow and
easier to find bugs in.  I left the flag name as-is for compatability reasons.

Fixes glamp/bashplotlib#1.